+++ Tests/Editor/MonoBehaviourEmptyMethodTests.cs
@@ -2,7 +2,7 @@ using System;
 using System.Linq;
 using NUnit.Framework;
 using Unity.ProjectAuditor.Editor;
-using Unity.ProjectAuditor.Editor.CodeAnalysis;
+using Unity.ProjectAuditor.Editor.AssemblyUtils;
 using Unity.ProjectAuditor.Editor.InstructionAnalyzers;
 using Unity.ProjectAuditor.Editor.Utils;
 
@@ -36,12 +36,12 @@ namespace Unity.ProjectAuditor.EditorTests
         [TestCase(CodeOptimization.Release)]
         public void CodeAnalysis_MonoBehaviourWithEmptyEventMethod_IsReported(CodeOptimization codeOptimization)
         {
-            var prevCodeOptimization = AssemblyCompilationPipeline.CodeOptimization;
-            AssemblyCompilationPipeline.CodeOptimization = codeOptimization;
+            var prevCodeOptimization = AssemblyCompilation.CodeOptimization;
+            AssemblyCompilation.CodeOptimization = codeOptimization;
 
             var scriptIssues = Utility.AnalyzeAndFindAssetIssues(m_MonoBehaviourWithEmptyEventMethod);
 
-            AssemblyCompilationPipeline.CodeOptimization = prevCodeOptimization; // restore previous value
+            AssemblyCompilation.CodeOptimization = prevCodeOptimization; // restore previous value
 
             Assert.AreEqual(1, scriptIssues.Count());