+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added SavePath to configuration asset
 * Added [Graphics Tier](https://docs.unity3d.com/ScriptReference/Rendering.GraphicsTier.html) information to reported Shader Variants
 * Added diagnostic message formatting support
+* Added dependencies panel to assembly view
+* Fixed reporting of assemblies not compiled due to dependencies
 * Improved code diagnostic messages
 
 ## [0.7.6-preview] - 2022-04-22